Unified cpp and cuda setup
