+++ wx_factory/common/transform_rule.hpp
@@ -1,0 +1,9 @@
+#ifndef TRANSFORM_RULE_H_
+#define TRANSFORM_RULE_H_
+
+struct TransformRule {
+    int s11, s12, s13, s14;
+    int s21, s22, s23, s24;
+};
+
+#endif
