+++ wx_factory/exchanges/kernels/kernels.h
@@ -2,6 +2,9 @@
 #define TEMPLATE_KERNELS_HPP
 
 #include "pairs.hpp"
+// #include "pairs_nogpu.hpp"
+
+// #include "flip_nogpu.hpp"
 #include "flip.hpp"
 
 #endif
